pyproject.toml
[tool.poetry]
name = "tomi-metric-collector"
version = "0.1.4"
description = ""
authors = ["Edgar"]
readme = "README.md"
packages = [{include = "tomi_metrics"}]

[tool.poetry.dependencies]
python = "^3.13"
flask = "^3.0.3"
gunicorn = "^23.0.0"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
pymongo = "^4.10.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
